feat(epics): calling `store.dispatch()` directly inside your epics is now deprecated and will be removed in v1.0.0 (#336)
Instead of using `store.dispatch()` directly, your epics should emit actions you wish to dispatch through the Observable you return. Actions in, actions out. `store.dispatch()` was made available as an escape hatch, but in practice it has been a footgun for many users. It will be removed in v1.0.0